Refactor internal naming
