+++ CHANGELOG.md
@@ -12,7 +12,8 @@
 - [x] Deprecate `createHook` for `runActions`
 - [x] Deprecate `registerAction({ hook: 'xxx' })` for `registerAction({ action: 'xxx', name: 'yyy' })`
 - [x] Deprecate `registerAction` for `registerExtension`
-- [ ] Deprecate `registerHook` for `registerActions` who always get a map of hooks to add
+- [x] Deprecate `registerHook` for `registerActions` who always get a map of hooks to add
+- [ ] Deprecate `getHook()` for `getAction()`
 - [ ] Fix kitchensink/mock-axios-e2e
 - [ ] Fix service order dependency in `service-fetchq` - it should work seamlessly before/after `service-fastify`
 - [ ] Should be able to `exports.name = 'featureName'; exports.actions = [{ hook, handler }, ...]`
